Share

+++ app/src/main/java/com/cgmdigitalhouse/cinelist/moviedetails/reviews/model/AuthorDetailsModel.kt
@@ -3,6 +3,6 @@ package com.cgmdigitalhouse.cinelist.moviedetails.reviews.model
 import com.google.gson.annotations.SerializedName
 
 data class AuthorDetailsModel (
-    @SerializedName("rating") val rating: Int,
+    @SerializedName("rating") val rating: Double,
     @SerializedName("username") val username: String
         )
